requirements-macos.txt
# macOS-specific requirements for EFIS Data Manager
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
watchdog>=2.1.0
paramiko>=2.11.0
cryptography>=37.0.0
pyyaml>=6.0
jsonschema>=4.0.0
pync>=2.0.3
plyer>=2.1.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991